depi-impl/depi/server/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "setuptools-grpc"]
build-backend = "setuptools.build_meta"

[project]
name = "depi-server"
version="0.7.1"
authors = [
    {name = "Mark Wutka", email="mark.wutka@vanderbilt.edu"},
    {name = "Patrik Meijer", email="patrik.meijer@vanderbilt.edu"}
]
requires-python = ">=3.10"
dependencies=[
    "build==1.0.3",
    "gitdb==4.0.10",
    "GitPython==3.1.31",
    "grpcio==1.63.0",
    "grpcio-tools==1.63.0",
    "protobuf==5.26.1",
    "pylint==2.17.5",
    "textX==3.1.1",
    "virtualenv==20.23.0",
    "waitress==2.1.2",
    "mysqlclient~=2.2.0"
]

[project.scripts]
depi-server="depi_server.depi_server:serve"
depi-config="depi_server.setup.get_config:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
depi_server = ["../depi_pb2.py", "../depi_pb2.pyi", "../depi_pb2_grpc.py",
    "auth/*.textx", "configs/depi_config_dolt.json", "configs/depi_config_mem.json", "configs/depi_config_mem_notls.json",
    "configs/depi_mysql.sql", "depi.proto"]
depi_pb2 = []
depi_pb2_grpc = []
